split out cms functions
