+++ app/api/util.ts
@@ -1,0 +1,7 @@
+export class CmsError extends Error {
+	data: any;
+	constructor(message: string, data?: any) {
+		super(message);
+		this.data = data;
+	}
+}
